Extracts classes from the motion module into their own files.
